refactor(Fixture) : 리팩토링 - Fixture 분리하기
