+++ src/test/java/kitchenpos/menus/tobe/domain/menu/MenuProductsTest.java
@@ -20,8 +20,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
-import static kitchenpos.Fixtures.INVALID_ID;
-import static kitchenpos.Fixtures.product;
+import static kitchenpos.fixture.Fixtures.INVALID_ID;
+import static kitchenpos.fixture.ProductFixture.product;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
