+++ src/test/java/kitchenpos/support/domain/OrderLineItemTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-import static kitchenpos.Fixtures.menu;
+import static kitchenpos.fixture.MenuFixture.menu;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
